Refactor: Phase 1 — fix profanity bug, split monolith, name constants
- Fix concurrency bug in CheckProfanity: was mutating goaway.DefaultProfanities
  global on every call AND the detector captured the list before mutation so
  filtering never actually worked. Replaced with ProfanityChecker struct
  initialized once with a filtered copy of the word list.

- Split 725-line prayertexter.go into 7 files by domain concern:
  prayertexter.go (routing), signup.go, prayer.go, completion.go,
  admin.go, member.go, helpers.go. No API or import path changes.

- Replace inline magic numbers with named constants:
  MinNameLetters, MinRequestWords, WeekDays, HoursPerDay,
  maxSendAttempts, sendRetryDelayMs, phoneRegexMatchCount.

Executed-By: mayor
